requirements.txt
kafka-python==2.0.2
redis==5.0.1
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.44.1
aiohttp==3.9.3
